refactor: test package move to domain
