+++ src/test/java/step3/domain/CarTest.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
