+++ src/test/java/step3/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package step3;
+package step3.domain;
 
 
 import org.junit.jupiter.api.BeforeEach;
